feat(core): do not rebuild native if prebuilt is present
According to documentation, this checks for an existing prebuild for the
current arch and skip building in `packages/core/build` if it found it.

"The install script will check if a compatible prebuild is bundled.
If so it does nothing. If not it will run node-gyp rebuild to produce a build.
This means that if the user using your module has disabled install scripts your
module will still work (!) as long as a compatible prebuild is bundled."

https://github.com/prebuild/prebuildify